requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
pyvista>=0.42.0
ultralytics>=8.0.0
tqdm>=4.65.0
filterpy>=1.4.5
scikit-image>=0.21.0
pyyaml>=6.0
matplotlib>=3.7.0
pillow>=9.0.0